# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GoodDollar Faucet API - Face verification, G$ token claiming, daily check-in streaks, and leaderboard on Celo
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ClaimInput {
  walletAddress: string;
  verificationProof: string;
  referralCode?: string | null;
}

export interface ClaimResult {
  success: boolean;
  txHash: string;
  amount: string;
  walletAddress: string;
  claimedAt: string;
}

export interface ClaimStatus {
  walletAddress: string;
  hasClaimed: boolean;
  isVerified: boolean;
  /** @nullable */
  claimedAt?: string | null;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  amount?: string | null;
}

export interface ClaimRecord {
  id: number;
  walletAddress: string;
  amount: string;
  txHash: string;
  claimedAt: string;
}

export interface FaucetStats {
  totalClaims: number;
  uniqueWallets: number;
  totalDistributed: string;
  claimsToday: number;
}

export interface VerificationStatus {
  walletAddress: string;
  isVerified: boolean;
  /** @nullable */
  verifiedAt: string | null;
}

export interface CheckinInput {
  walletAddress: string;
}

export interface CheckinResult {
  success: boolean;
  currentStreak: number;
  bonusEarned: number;
  milestoneReached: boolean;
  /** @nullable */
  milestoneName?: string | null;
  nextMilestone: number;
  message: string;
  totalBonusEarned?: number;
}

export interface MilestoneInfo {
  day: number;
  bonus: number;
  label: string;
  reached: boolean;
}

export interface StreakStatus {
  walletAddress: string;
  currentStreak: number;
  longestStreak: number;
  totalCheckins: number;
  totalBonusEarned: number;
  canCheckinToday: boolean;
  /** @nullable */
  lastCheckinAt?: string | null;
  nextMilestone: number;
  currentCycleDay: number;
  milestones: MilestoneInfo[];
}

export interface LeaderboardEntry {
  rank: number;
  walletAddress: string;
  currentStreak: number;
  longestStreak: number;
  totalCheckins: number;
  totalBonusEarned: number;
}

export interface ReferralRecord {
  referredWallet: string;
  bonusEarned: number;
  createdAt: string;
}

export interface ReferralStats {
  walletAddress: string;
  referralCode: string;
  referralLink: string;
  totalReferrals: number;
  bonusEarned: number;
  referrals: ReferralRecord[];
}

export interface ReferralValidation {
  valid: boolean;
  referralCode: string;
  /** @nullable */
  referrerWallet: string | null;
}

export interface ErrorResponse {
  error: string;
}