# lib/api-zod/src/generated/types/claimResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GoodDollar Faucet API - Face verification, G$ token claiming, daily check-in streaks, and leaderboard on Celo
 * OpenAPI spec version: 0.1.0
 */

export interface ClaimResult {
  success: boolean;
  txHash: string;
  amount: string;
  walletAddress: string;
  claimedAt: Date;
}