# lib/api-zod/src/generated/types/streakStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GoodDollar Faucet API - Face verification, G$ token claiming, daily check-in streaks, and leaderboard on Celo
 * OpenAPI spec version: 0.1.0
 */
import type { MilestoneInfo } from './milestoneInfo';

export interface StreakStatus {
  walletAddress: string;
  currentStreak: number;
  longestStreak: number;
  totalCheckins: number;
  totalBonusEarned: number;
  canCheckinToday: boolean;
  /** @nullable */
  lastCheckinAt?: Date | null;
  nextMilestone: number;
  currentCycleDay: number;
  milestones: MilestoneInfo[];
}